src/planning_layer/recipe_planner.py
"""
Recipe Planner - LLM-powered task decomposition
Converts high-level recipe instructions into structured task sequences.
"""

from typing import List, Dict, Any
import yaml
from pathlib import Path

class RecipePlanner:
    """Plans recipe execution using LLM-based task decomposition."""
    
    def __init__(self, config_path: str = "config/models/llm_config.yaml"):
        self.config = self._load_config(config_path)
        # TODO: Initialize LLM client
        
    def _load_config(self, config_path: str) -> Dict[str, Any]:
        """Load planner configuration."""
        config_file = Path(config_path)
        if config_file.exists():
            with open(config_file, 'r') as f:
                return yaml.safe_load(f)
        return {"model": "qwen-2.5", "endpoint": "http://localhost:8001/v1"}
    
    def plan_task(self, instruction: str) -> List[Dict[str, Any]]:
        """
        Plan a cooking task from natural language instruction.
        
        Args:
            instruction: Natural language cooking instruction
            
        Returns:
            List of structured task steps
        """
        # TODO: Implement LLM-based planning
        # For now, return a placeholder
        if "carrot" in instruction.lower() and "slicer" in instruction.lower():
            return [
                {
                    "action": "pick_and_place",
                    "object": "carrot", 
                    "source": "counter",
                    "target": "slicer",
                    "priority": 1
                }
            ]
        
        return []
    
    def load_recipe(self, recipe_name: str) -> Dict[str, Any]:
        """Load recipe configuration from file."""
        recipe_path = Path(f"config/recipes/{recipe_name}.yaml")
        with open(recipe_path, 'r') as f:
            return yaml.safe_load(f)
