tests/integration/test_full_pipeline/test_carrot_pickup.py
"""
Phase 1 Integration Test: Carrot Pickup Task
Tests the complete pipeline from planning to execution for basic pickup task.
"""

import pytest
import asyncio
from unittest.mock import Mock, AsyncMock

from src.planning_layer.recipe_planner import RecipePlanner
from src.movement_layer.grasp_planner import GraspPlanner
from src.control_layer.motion_planner import MotionPlanner


@pytest.fixture
def mock_components():
    """Create mock components for testing."""
    return {
        'planner': Mock(spec=RecipePlanner),
        'grasp_planner': Mock(spec=GraspPlanner), 
        'motion_planner': Mock(spec=MotionPlanner),
    }


@pytest.mark.asyncio
async def test_carrot_pickup_planning(mock_components):
    """Test that planning layer can decompose carrot pickup task."""
    planner = mock_components['planner']
    
    # Mock the planning response
    expected_steps = [
        {"action": "pick_and_place", "object": "carrot", "source": "counter", "target": "slicer"},
    ]
    planner.plan_task.return_value = expected_steps
    
    # Execute planning
    result = planner.plan_task("prepare steamed carrots - step 1: move carrot to slicer")
    
    # Verify planning output
    assert len(result) == 1
    assert result[0]["action"] == "pick_and_place"
    assert result[0]["object"] == "carrot"


@pytest.mark.asyncio 
async def test_carrot_grasp_planning(mock_components):
    """Test that movement layer can plan carrot grasping."""
    grasp_planner = mock_components['grasp_planner']
    
    # Mock grasp pose response
    expected_pose = {
        "position": {"x": 0.3, "y": 0.1, "z": 0.15},
        "orientation": {"rx": 0, "ry": 0, "rz": 0},
        "gripper": "open"
    }
    grasp_planner.plan_grasp.return_value = expected_pose
    
    # Execute grasp planning
    mock_camera_data = {"rgb": "mock_image", "depth": "mock_depth"}
    result = grasp_planner.plan_grasp(mock_camera_data, "carrot")
    
    # Verify grasp planning
    assert "position" in result
    assert "orientation" in result
    assert result["gripper"] == "open"


@pytest.mark.asyncio
async def test_motion_execution(mock_components):
    """Test that control layer can execute motion."""
    motion_planner = mock_components['motion_planner']
    
    # Mock successful motion execution
    motion_planner.execute_pose.return_value = {"success": True, "message": "Motion completed"}
    
    # Execute motion
    target_pose = {
        "position": {"x": 0.3, "y": 0.1, "z": 0.15},
        "orientation": {"rx": 0, "ry": 0, "rz": 0}
    }
    result = motion_planner.execute_pose(target_pose)
    
    # Verify execution
    assert result["success"] is True


@pytest.mark.asyncio
async def test_full_carrot_pickup_pipeline(mock_components):
    """Integration test for complete carrot pickup pipeline."""
    
    # This test will be implemented once the actual components are built
    # For now, it serves as a placeholder and specification
    
    pipeline_steps = [
        "1. Plan task: 'move carrot to slicer'",
        "2. Analyze camera feed to locate carrot", 
        "3. Plan grasp approach for carrot",
        "4. Execute approach motion",
        "5. Close gripper on carrot",
        "6. Plan placement motion to slicer",
        "7. Execute placement motion", 
        "8. Open gripper to release carrot"
    ]
    
    # Verify our test covers the essential pipeline
    assert len(pipeline_steps) == 8
    
    # TODO: Implement actual integration test once components are ready
    pytest.skip("Full pipeline test pending component implementation")


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
