requirements/base.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pyyaml>=5.4.0
requests>=2.25.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
loguru>=0.5.0
pillow>=8.0.0
scipy>=1.7.0
matplotlib>=3.5.0
